Add switch to $patch_levels for Magento 2 Official Patches
Magento is a fairly popular software where composer-patches is especially useful.  Patch files distributed by Magento in some special cases are based in the root directory, containing paths like `a/vendor/magento/module-catalog`.  In cases such as these, if only a single module is referenced in the patch file then a -p4 flag will allow the patch to work as distributed by Magento.