Communicate that patch failed due to missing 'patch' command
## Problem
When attempting to use this package on a system that does not have the "patch" command installed, you get a vague error message:

> Could not apply patch! Skipping. The error was: Cannot apply patch ######

## Cause
I spend a long time trying to work out exactly why my patch was not applying. Thinking that it was a problem with the patch file itself. But eventually, I discovered that the package is assuming the system has a "patch" command.

##
Check if the patch command is present and can be run before trying to use it. If that fails, then clearly communicate to the user what the problem is so they can resolve it.